modify shell script

+++ install.sh
@@ -1,9 +1,21 @@
-sudo apt update
-
 RED='\033[0;31m'
 NC='\033[0m'
+YELLOW='\033[1;33m'
+BLUE='\033[0;34m'
 
-echo -e ${RED}start installing node${NC}
+echo -e ${RED}Warning this script will install vim, vim -plug, node, python, ripgrep, code-minimap and gcc, do you still want to continue?'(Y/n)'${NC} 
+
+read shouldContinue
+
+if [ $shouldContinue != 'y' ]
+then 
+	exit
+fi
+
+sudo apt update
+
+#install v14.17.5 node
+echo -e ${BLUE}start installing node${NC}
 sudo apt remove -y nodejs
 NODEAMD64='node-v14.17.5-linux-x64'
 NODEAARCH64='node-v14.17.5-linux-arm64'
@@ -27,27 +39,31 @@ case $(uname -m) in
 	*) installNode $NODEAMD64
 		;;
 esac
-echo -e ${RED}finish${NC}
+echo -e ${BLUE}finish${NC}
 
-echo -e  ${RED}start installing vim${NC}
+#install the newest vim
+echo -e  ${BLUE}start installing vim${NC}
 sudo apt install -y libncurses-dev make gcc python3 python
 sudo apt remove -y vim
 sudo rm /usr/local/bin/*vim*
 sudo rm /usr/bin/*vim*
 git clone https://github.com/vim/vim.git && cd vim
-echo -e  ${RED}configuring the vim feature${NC}
+echo -e  ${YELLOW}configuring the vim feature${NC}
 ./configure --enable-pythoninterp=yes --enable-python3interp=yes --prefix=/usr/local/
+echo -e ${YELLOW}finish${NC}
 make && sudo make install
 cd ..
 rm -rf vim
 source ~/.bashrc
-echo -e ${RED}finish${NC}
+echo -e ${BLUE}finish${NC}
 
-echo -e  ${RED}start installing ripgrep${NC}
+#install ripgrep for fzf plugin
+echo -e  ${BLUE}start installing ripgrep${NC}
 sudo apt install -y ripgrep
-echo -e ${RED}finish${NC}
+echo -e ${BLUE}finish${NC}
 
-echo -e ${RED}start installing minimap${NC}
+#install minimap for minimap plugin
+echo -e ${BLUE}start installing minimap${NC}
 MINIMAPAMD64='code-minimap-v0.6.1-i686-unknown-linux-musl'
 MINIMAPAARCH64='code-minimap-v0.6.1-aarch64-unknown-linux-gnu'
 MINIMAPARMV7='code-minimap-v0.6.1-arm-unknown-linux-gnueabihf'
@@ -69,19 +85,21 @@ case $(uname -m) in
 	*) installMiniMap $MINIMAPAMD64
 		;;
 esac
-echo -e ${RED}finish${NC}
+echo -e ${BLUE}finish${NC}
 
-echo -e ${RED}start setting up vimrc${NC}
+#install vim-plug and set up vimrc
+echo -e ${BLUE}start setting up vimrc${NC}
 sudo cat my.vim > ~/.vimrc
 curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
     https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
 vim +PlugInstall +qa
-echo -e ${RED}finish${NC}
+echo -e ${BLUE}finish${NC}
 
-echo -e ${RED}start setting CocConfig${NC}
+#set up coc configs
+echo -e ${BLUE}start setting up coc${NC}
 sudo cp ./coc-settings.json ~/.vim/
 sudo cp -r ./ultisnips ~/.config/coc/
-echo -e ${RED}finish${NC}
+echo -e ${BLUE}finish${NC}
 
-echo -e ${RED}success${NC}
+echo -e ${BLUE}success${NC}
